Fix/GitHub tests

+++ test/helpers/MockAggregatorV3.sol
@@ -1,5 +1,6 @@
 pragma solidity >=0.6.12;
 pragma experimental ABIEncoderV2;
+
 import {DSThing} from "ds-thing/thing.sol";
 
 interface AggregatorV3Interface {
@@ -9,9 +10,7 @@ interface AggregatorV3Interface {
 
     function version() external view returns (uint256);
 
-    function getRoundData(
-        uint80 _roundId
-    )
+    function getRoundData(uint80 _roundId)
         external
         view
         returns (uint80 roundId, int256 answer, uint256 startedAt, uint256 updatedAt, uint80 answeredInRound);
@@ -24,10 +23,12 @@ interface AggregatorV3Interface {
 
 contract MockAggregatorV3 is AggregatorV3Interface, DSThing {
     // --- Auth ---
-    mapping(address => uint) public wards;
+    mapping(address => uint256) public wards;
+
     function rely(address guy) external auth {
         wards[guy] = 1;
     }
+
     function deny(address guy) external auth {
         wards[guy] = 0;
     }
@@ -37,7 +38,7 @@ contract MockAggregatorV3 is AggregatorV3Interface, DSThing {
     uint8 public override decimals = 6;
 
     int256 internal answer = 0;
-    uint internal live = 0;
+    uint256 internal live = 0;
 
     // --- Init ---
     constructor() public {
@@ -46,20 +47,19 @@ contract MockAggregatorV3 is AggregatorV3Interface, DSThing {
     }
 
     // --- Administration ---
-    function file(bytes32 what, uint data) external auth {
+    function file(bytes32 what, uint256 data) external auth {
         require(live == 1, "MockAggregatorV3/not-live");
         if (what == "decimals") decimals = uint8(data);
         else revert("MockAggregatorV3/file-unrecognized-param");
     }
+
     function file(bytes32 what, int256 data) external auth {
         require(live == 1, "MockAggregatorV3/not-live");
         if (what == "answer") answer = data;
         else revert("MockAggregatorV3/file-unrecognized-param");
     }
 
-    function getRoundData(
-        uint80 _roundId
-    )
+    function getRoundData(uint80 _roundId)
         external
         view
         override
